+++ src/Config/ConfigOutputFactory.php
@@ -10,7 +10,7 @@ use Yiisoft\Composer\Config\Builder;
  * Config factory creates Config object of proper class
  * according to config name (and maybe other options later).
  */
-class ConfigFactory
+class ConfigOutputFactory
 {
     private const KNOWN_TYPES = [
         '__files' => System::class,
@@ -20,9 +20,9 @@ class ConfigFactory
         'constants' => Constants::class,
     ];
 
-    public function create(Builder $builder, string $name): Config
+    public function create(Builder $builder, string $name): ConfigOutput
     {
-        $class = self::KNOWN_TYPES[$name] ?? Config::class;
+        $class = self::KNOWN_TYPES[$name] ?? ConfigOutput::class;
 
         return new $class($builder, $name);
     }